Added User input
Used a Keyboard Module found at https://pypi.org/project/keyboard/ to
implement user control. Also added in a top function that only takes a
string as an input, and outputs the List of Lists containing the
permutations of the original string.

The program prompts the user to enter their letters, and then prints out
all permutations of those letters. Then, the program prompts the user to
either enter more letters, or pres Esc to exit the program.